Refactor parser to support multiple blanks, empty lines. (#4)
The Matrix Market spec allows any number of spaces before/between/after fields
or empty lines, so we want to be able to handle those. Conversely, the
%%MatrixMarket banner MUST be at the start, so we enforce that in the spec.

As part of the refactoring, we simplify the state tracking during parsing, by
removing the 'onto' field variable and reducing the number of intermediate
strings. This requires a bit more hard-coding per object/format type.

Vector objects may now omit the symmetry field; this is set it to GENERAL,
because none of the others really apply for vectors. Any number of additional
specifiers in the banner line is allowed and will be ignored.

Greatly fleshed out the tests to make sure we cover all of the new code. We
also spice up our simulations with variable spaces and comments.