Represent indices and line counts using a guaranteed >=64-bit integer.
We don't use size_t by default as this might not be big enough - the standard
only guarantees 16 bits, and 32-bit implementations might only use 32 bits,
which would lead to overflows when dealing with more than 4 billion lines.

We replace size_t's with an implicit size_type when iterating over vectors in
the tests. We limit our use of size_t's to C-style array indexing, and we slap
a std:: prefix to ensure that we are compliant with the spec.