Parallelize the parsing by processing chunks of newlines. (#5)
Specifically, we split the input stream into approximately equisized
newline-terminated blocks, then distribute each block to a worker in a
threadpool for parsing; once parsed, we then call the user-defined store()
function in a serial section. This is faster than using PerByteParallel as we
can make better use of all available threads.