Switch back to std::stod for accurate float parsing.
Our custom code wasn't quite exact and this complicates downstream usage.

That said, std::stod and friends aren't great either, as they depend on the
application's locale; I don't have a safe way to set and restore the locale.

This switch is just a placeholder for the promised day that we can use
std::from_chars. To that end, we explicitly prohibit hexadecimal floats.

Incidentally, we also force the Type_ of scan_integer to be integer, and the
Type_ of scan_real/scan_complex to be floating-point, to avoid trouble.